+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.test.ts
@@ -2,6 +2,7 @@ import { ControllerMessenger } from '@metamask/base-controller';
 import type { InternalAccount } from '@metamask/keyring-api';
 import type nock from 'nock';
 
+import encryption, { createSHA256Hash } from '../../shared/encryption';
 import type {
   AuthenticationControllerGetBearerToken,
   AuthenticationControllerGetSessionProfile,
@@ -13,6 +14,7 @@ import {
   MOCK_USER_STORAGE_ACCOUNTS,
 } from './__fixtures__/mockAccounts';
 import {
+  mockEndpointBatchUpsertUserStorage,
   mockEndpointGetUserStorage,
   mockEndpointGetUserStorageAllFeatureEntries,
   mockEndpointUpsertUserStorage,
@@ -23,7 +25,6 @@ import {
   MOCK_STORAGE_KEY_SIGNATURE,
 } from './__fixtures__/mockStorage';
 import * as AccountsUserStorageModule from './accounts/user-storage';
-import encryption from './encryption/encryption';
 import type {
   GetUserStorageAllFeatureEntriesResponse,
   GetUserStorageResponse,
@@ -504,12 +505,27 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
               'accounts',
               mockUserStorageAccountsResponse,
             ),
-          mockEndpointUpsertUserStorageAccount1: mockEndpointUpsertUserStorage(
-            `accounts.${MOCK_INTERNAL_ACCOUNTS.ALL[0].address}`,
-          ),
-          mockEndpointUpsertUserStorageAccount2: mockEndpointUpsertUserStorage(
-            `accounts.${MOCK_INTERNAL_ACCOUNTS.ALL[1].address}`,
-          ),
+          mockEndpointBatchUpsertUserStorage:
+            mockEndpointBatchUpsertUserStorage(
+              'accounts',
+              undefined,
+              async (_uri, requestBody) => {
+                const decryptedBody = await decryptBatchUpsertBody(
+                  requestBody,
+                  MOCK_STORAGE_KEY,
+                );
+
+                const expectedBody = createExpectedAccountSyncBatchUpsertBody(
+                  MOCK_INTERNAL_ACCOUNTS.ALL.slice(0, 2).map((account) => [
+                    account.address,
+                    account as InternalAccount,
+                  ]),
+                  MOCK_STORAGE_KEY,
+                );
+
+                expect(decryptedBody).toStrictEqual(expectedBody);
+              },
+            ),
         },
       };
     };
@@ -525,12 +541,9 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
 
     await controller.syncInternalAccountsWithUserStorage();
     mockAPI.mockEndpointGetUserStorage.done();
-    mockAPI.mockEndpointUpsertUserStorageAccount1.done();
-    mockAPI.mockEndpointUpsertUserStorageAccount2.done();
 
     expect(mockAPI.mockEndpointGetUserStorage.isDone()).toBe(true);
-    expect(mockAPI.mockEndpointUpsertUserStorageAccount1.isDone()).toBe(true);
-    expect(mockAPI.mockEndpointUpsertUserStorageAccount2.isDone()).toBe(true);
+    expect(mockAPI.mockEndpointBatchUpsertUserStorage.isDone()).toBe(true);
   });
 
   it('creates internal accounts if user storage has more accounts', async () => {
@@ -660,10 +673,8 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
               'accounts',
               await mockUserStorageAccountsResponse(),
             ),
-
-          mockEndpointUpsertUserStorageAccount2: mockEndpointUpsertUserStorage(
-            `accounts.${MOCK_INTERNAL_ACCOUNTS.ALL[1].address}`,
-          ),
+          mockEndpointBatchUpsertUserStorage:
+            mockEndpointBatchUpsertUserStorage('accounts'),
         },
       };
     };
@@ -680,9 +691,10 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
     await controller.syncInternalAccountsWithUserStorage();
 
     mockAPI.mockEndpointGetUserStorage.done();
-    mockAPI.mockEndpointUpsertUserStorageAccount2.done();
+    mockAPI.mockEndpointBatchUpsertUserStorage.done();
 
     expect(mockAPI.mockEndpointGetUserStorage.isDone()).toBe(true);
+    expect(mockAPI.mockEndpointBatchUpsertUserStorage.isDone()).toBe(true);
 
     expect(messengerMocks.mockKeyringAddNewAccount).not.toHaveBeenCalled();
   });
@@ -749,9 +761,8 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
                 'accounts',
                 await mockUserStorageAccountsResponse(),
               ),
-            mockEndpointUpsertUserStorage: mockEndpointUpsertUserStorage(
-              `accounts.${MOCK_INTERNAL_ACCOUNTS.ONE_CUSTOM_NAME_WITHOUT_LAST_UPDATED[0].address}`,
-            ),
+            mockEndpointBatchUpsertUserStorage:
+              mockEndpointBatchUpsertUserStorage('accounts'),
           },
         };
       };
@@ -768,7 +779,7 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
       await controller.syncInternalAccountsWithUserStorage();
 
       mockAPI.mockEndpointGetUserStorage.done();
-      mockAPI.mockEndpointUpsertUserStorage.done();
+      mockAPI.mockEndpointBatchUpsertUserStorage.done();
 
       expect(
         messengerMocks.mockAccountsUpdateAccountMetadata,
@@ -790,9 +801,8 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
                 'accounts',
                 await mockUserStorageAccountsResponse(),
               ),
-            mockEndpointUpsertUserStorage: mockEndpointUpsertUserStorage(
-              `accounts.${MOCK_INTERNAL_ACCOUNTS.ONE_CUSTOM_NAME_WITH_LAST_UPDATED[0].address}`,
-            ),
+            mockEndpointBatchUpsertUserStorage:
+              mockEndpointBatchUpsertUserStorage('accounts'),
           },
         };
       };
@@ -809,7 +819,7 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
       await controller.syncInternalAccountsWithUserStorage();
 
       mockAPI.mockEndpointGetUserStorage.done();
-      mockAPI.mockEndpointUpsertUserStorage.done();
+      mockAPI.mockEndpointBatchUpsertUserStorage.done();
 
       expect(
         messengerMocks.mockAccountsUpdateAccountMetadata,
@@ -922,9 +932,8 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
                 'accounts',
                 await mockUserStorageAccountsResponse(),
               ),
-            mockEndpointUpsertUserStorage: mockEndpointUpsertUserStorage(
-              `accounts.${MOCK_INTERNAL_ACCOUNTS.ONE_CUSTOM_NAME_WITH_LAST_UPDATED[0].address}`,
-            ),
+            mockEndpointBatchUpsertUserStorage:
+              mockEndpointBatchUpsertUserStorage('accounts'),
           },
         };
       };
@@ -941,7 +950,7 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
       await controller.syncInternalAccountsWithUserStorage();
 
       mockAPI.mockEndpointGetUserStorage.done();
-      mockAPI.mockEndpointUpsertUserStorage.done();
+      mockAPI.mockEndpointBatchUpsertUserStorage.done();
 
       expect(
         messengerMocks.mockAccountsUpdateAccountMetadata,
@@ -1150,9 +1159,8 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
                 'accounts',
                 await mockUserStorageAccountsResponse(),
               ),
-            mockEndpointUpsertUserStorage: mockEndpointUpsertUserStorage(
-              `accounts.${MOCK_INTERNAL_ACCOUNTS.ONE_CUSTOM_NAME_WITH_LAST_UPDATED_MOST_RECENT[0].address}`,
-            ),
+            mockEndpointBatchUpsertUserStorage:
+              mockEndpointBatchUpsertUserStorage('accounts'),
           },
         };
       };
@@ -1169,7 +1177,7 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
       await controller.syncInternalAccountsWithUserStorage();
 
       mockAPI.mockEndpointGetUserStorage.done();
-      mockAPI.mockEndpointUpsertUserStorage.done();
+      mockAPI.mockEndpointBatchUpsertUserStorage.done();
 
       expect(
         messengerMocks.mockAccountsUpdateAccountMetadata,
@@ -1425,7 +1433,13 @@ function mockUserStorageMessenger(options?: {
       NotificationServicesControllerDisableNotificationServices['handler']
     >().mockResolvedValue();
 
-  const mockKeyringAddNewAccount = jest.fn().mockResolvedValue('0x123');
+  const mockKeyringAddNewAccount = jest.fn(() => {
+    baseMessenger.publish(
+      'AccountsController:accountAdded',
+      MOCK_INTERNAL_ACCOUNTS.ONE[0] as InternalAccount,
+    );
+    return MOCK_INTERNAL_ACCOUNTS.ONE[0].address;
+  });
 
   const mockAccountsListAccounts = jest
     .fn()
@@ -1561,4 +1575,49 @@ async function createMockUserStorageEntries(
   data: unknown[],
 ): Promise<GetUserStorageAllFeatureEntriesResponse> {
   return await Promise.all(data.map((d) => createMockUserStorageEntry(d)));
+}
+
+/**
+ * Test Utility - decrypts a realistic batch upsert payload
+ * @param requestBody - nock body
+ * @param storageKey - storage key
+ * @returns decrypted body
+ */
+async function decryptBatchUpsertBody(
+  requestBody: nock.Body,
+  storageKey: string,
+) {
+  if (typeof requestBody === 'string') {
+    return requestBody;
+  }
+  return await Promise.all(
+    Object.entries<string>(requestBody.data).map(
+      async ([entryKey, entryValue]) => {
+        return [
+          entryKey,
+          await encryption.decryptString(entryValue, storageKey),
+        ];
+      },
+    ),
+  );
+}
+
+/**
+ * Test Utility - creates a realistic expected batch upsert payload
+ * @param data - data supposed to be upserted
+ * @param storageKey - storage key
+ * @returns expected body
+ */
+function createExpectedAccountSyncBatchUpsertBody(
+  data: [string, InternalAccount][],
+  storageKey: string,
+) {
+  return data.map(([entryKey, entryValue]) => [
+    createSHA256Hash(String(entryKey) + storageKey),
+    JSON.stringify(
+      AccountsUserStorageModule.mapInternalAccountToUserStorageAccount(
+        entryValue,
+      ),
+    ),
+  ]);
 }