+++ packages/profile-sync-controller/src/sdk/user-storage.test.ts
@@ -1,3 +1,7 @@
+import type { UserStoragePathWithKeyOnly } from 'src/shared/storage-schema';
+
+import encryption, { createSHA256Hash } from '../shared/encryption';
+import { Env } from '../shared/env';
 import { arrangeAuthAPIs } from './__fixtures__/mock-auth';
 import {
   MOCK_NOTIFICATIONS_DATA,
@@ -8,7 +12,6 @@ import {
 } from './__fixtures__/mock-userstorage';
 import { arrangeAuth, typedMockFn } from './__fixtures__/test-utils';
 import type { IBaseAuth } from './authentication-jwt-bearer/types';
-import { Env } from './env';
 import { NotFoundError, UserStorageError } from './errors';
 import type { StorageOptions } from './user-storage';
 import { STORAGE_URL, UserStorage } from './user-storage';
@@ -89,6 +92,46 @@ describe('User Storage', () => {
     expect(responseAllFeatureEntries).toStrictEqual([data]);
   });
 
+  it('batch set items', async () => {
+    const dataToStore: [UserStoragePathWithKeyOnly, string][] = [
+      ['0x123', JSON.stringify(MOCK_NOTIFICATIONS_DATA)],
+      ['0x456', JSON.stringify(MOCK_NOTIFICATIONS_DATA)],
+    ];
+
+    const { auth } = arrangeAuth('SRP', MOCK_SRP);
+    const { userStorage } = arrangeUserStorage(auth);
+
+    const mockPut = handleMockUserStoragePut(
+      undefined,
+      async (_, requestBody) => {
+        if (typeof requestBody === 'string') {
+          return;
+        }
+
+        const decryptedBody = await Promise.all(
+          Object.entries<string>(requestBody.data).map(
+            async ([entryKey, entryValue]) => {
+              return [
+                entryKey,
+                await encryption.decryptString(entryValue, MOCK_STORAGE_KEY),
+              ];
+            },
+          ),
+        );
+
+        const expectedBody = dataToStore.map(([entryKey, entryValue]) => [
+          createSHA256Hash(String(entryKey) + MOCK_STORAGE_KEY),
+          entryValue,
+        ]);
+
+        expect(decryptedBody).toStrictEqual(expectedBody);
+      },
+    );
+
+    await userStorage.batchSetItems('accounts', dataToStore);
+    expect(mockPut.isDone()).toBe(true);
+  });
+
   it('user storage: failed to set key', async () => {
     const { auth } = arrangeAuth('SRP', MOCK_SRP);
     const { userStorage } = arrangeUserStorage(auth);
@@ -107,6 +150,23 @@ describe('User Storage', () => {
     ).rejects.toThrow(UserStorageError);
   });
 
+  it('user storage: failed to batch set items', async () => {
+    const { auth } = arrangeAuth('SRP', MOCK_SRP);
+    const { userStorage } = arrangeUserStorage(auth);
+
+    handleMockUserStoragePut({
+      status: 401,
+      body: {
+        message: 'failed to insert storage entries',
+        error: 'generic-error',
+      },
+    });
+
+    await expect(
+      userStorage.batchSetItems('notifications', [['key', 'value']]),
+    ).rejects.toThrow(UserStorageError);
+  });
+
   it('user storage: failed to get storage entry', async () => {
     const { auth } = arrangeAuth('SRP', MOCK_SRP);
     const { userStorage } = arrangeUserStorage(auth);
@@ -121,6 +181,23 @@ describe('User Storage', () => {
 
     await expect(
       userStorage.getItem('notifications.notification_settings'),
+    ).rejects.toThrow(UserStorageError);
+  });
+
+  it('user storage: failed to get storage entries', async () => {
+    const { auth } = arrangeAuth('SRP', MOCK_SRP);
+    const { userStorage } = arrangeUserStorage(auth);
+
+    await handleMockUserStorageGetAllFeatureEntries({
+      status: 401,
+      body: {
+        message: 'failed to get storage entries',
+        error: 'generic-error',
+      },
+    });
+
+    await expect(
+      userStorage.getAllFeatureItems('notifications'),
     ).rejects.toThrow(UserStorageError);
   });