+++ packages/profile-sync-controller/src/sdk/user-storage.ts
@@ -1,13 +1,13 @@
+import encryption, { createSHA256Hash } from '../shared/encryption';
+import type { Env } from '../shared/env';
+import { getEnvUrls } from '../shared/env';
 import type {
   UserStoragePathWithFeatureAndKey,
   UserStoragePathWithFeatureOnly,
-} from '../controllers/user-storage/schema';
-import { createEntryPath } from '../controllers/user-storage/schema';
-import type { GetUserStorageAllFeatureEntriesResponse } from '../controllers/user-storage/services';
+  UserStoragePathWithKeyOnly,
+} from '../shared/storage-schema';
+import { createEntryPath } from '../shared/storage-schema';
 import type { IBaseAuth } from './authentication-jwt-bearer/types';
-import encryption, { createSHA256Hash } from './encryption';
-import type { Env } from './env';
-import { getEnvUrls } from './env';
 import { NotFoundError, UserStorageError } from './errors';
 
 export const STORAGE_URL = (env: Env, encryptedPath: string) =>
@@ -26,6 +26,13 @@ export type StorageOptions = {
 export type UserStorageOptions = {
   storage?: StorageOptions;
 };
+
+type GetUserStorageAllFeatureEntriesResponse = {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  HashedKey: string;
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Data: string;
+}[];
 
 type ErrorMessage = {
   message: string;
@@ -50,6 +57,13 @@ export class UserStorage {
     value: string,
   ): Promise<void> {
     await this.#upsertUserStorage(path, value);
+  }
+
+  async batchSetItems(
+    path: UserStoragePathWithFeatureOnly,
+    values: [UserStoragePathWithKeyOnly, string][],
+  ) {
+    await this.#batchUpsertUserStorage(path, values);
   }
 
   async getItem(path: UserStoragePathWithFeatureAndKey): Promise<string> {
@@ -113,6 +127,57 @@ export class UserStorage {
         e instanceof Error ? e.message : JSON.stringify(e ?? '');
       throw new UserStorageError(
         `failed to upsert user storage for path '${path}'. ${errorMessage}`,
+      );
+    }
+  }
+
+  async #batchUpsertUserStorage(
+    path: UserStoragePathWithFeatureOnly,
+    data: [UserStoragePathWithKeyOnly, string][],
+  ): Promise<void> {
+    try {
+      if (!data.length) {
+        return;
+      }
+
+      const headers = await this.#getAuthorizationHeader();
+      const storageKey = await this.getStorageKey();
+
+      const encryptedData = await Promise.all(
+        data.map(async (d) => {
+          return [
+            this.#createEntryKey(d[0], storageKey),
+            await encryption.encryptString(d[1], storageKey),
+          ];
+        }),
+      );
+
+      const url = new URL(STORAGE_URL(this.env, path));
+
+      const response = await fetch(url.toString(), {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+          ...headers,
+        },
+        body: JSON.stringify({ data: Object.fromEntries(encryptedData) }),
+      });
+
+      if (!response.ok) {
+        const responseBody: ErrorMessage = await response.json().catch(() => ({
+          message: 'unknown',
+          error: 'unknown',
+        }));
+        throw new Error(
+          `HTTP error message: ${responseBody.message}, error: ${responseBody.error}`,
+        );
+      }
+    } catch (e) {
+      /* istanbul ignore next */
+      const errorMessage =
+        e instanceof Error ? e.message : JSON.stringify(e ?? '');
+      throw new UserStorageError(
+        `failed to batch upsert user storage for path '${path}'. ${errorMessage}`,
       );
     }
   }